+++ tests/HasRolesTest.php
@@ -378,6 +378,23 @@ class HasRolesTest extends TestCase
 		$this->assertTrue($this->testUser->hasAnyPermission('edit-news', 'edit-articles'));
 	}
 
+    /** @test */
+    public function it_can_determine_that_user_has_direct_permission()
+    {
+        $this->testUser->givePermissionTo('edit-articles');
+        $this->refreshTestUser();
+        $this->assertTrue($this->testUser->hasDirectPermission('edit-articles'));
+        $this->testUser->revokePermissionTo('edit-articles');
+        $this->refreshTestUser();
+        $this->assertFalse($this->testUser->hasDirectPermission('edit-articles'));
+
+        $this->testUser->assignRole('testRole');
+        $this->testRole->givePermissionTo('edit-articles');
+        $this->refreshTestUser();
+        $this->assertFalse($this->testUser->hasDirectPermission('edit-articles'));
+    }
+
+
     /**
      * @test
      *